requirements.txt
# fastapi==0.104.1
# uvicorn==0.24.0
# python-multipart==0.0.6
# pydantic==2.4.2
# pydantic-settings==2.1.0
# python-jose==3.3.0
# passlib==1.7.4
# python-dotenv==1.0.0
# requests==2.31.0
# aiohttp==3.9.1
# pytest==7.4.3
# httpx==0.25.1
# pytest-asyncio==0.21.1
# PyPDF2==3.0.1
# openai==1.12.0
# # firecrawl-py==0.0.17 # No longer using Firecrawl

fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.4.2
pytest==7.4.3
python-dotenv==1.0.0
pydantic-settings==2.1.0
openai==1.3.5
celery==5.3.6
redis==5.0.4
httpx==0.25.1
tenacity==8.2.1
beautifulsoup4==4.12.3
lxml==5.2.2 